+++ scripts/testng-junit/src/tests/setup.py
@@ -2,8 +2,7 @@ import os
 import sys
 sys.path.insert(0, os.path.abspath(os.path.join(os.path.dirname(__file__), '..')))
 
-import testng2junit
-
+import testng2junit4
 
 ###
 # Strip content for equality comparison.
