+++ scripts/testng-junit/src/tests/test_migrate_exception.py
@@ -1,4 +1,4 @@
-from setup import testng2junit
+from setup import testng2junit4
 from setup import assert_equal_content
 
 content = """
@@ -113,5 +113,5 @@ expected = """
 
 
 def test_migrate_exceptions():
-    content_new = testng2junit.migrate_exceptions(content)
+    content_new = testng2junit4.migrate_exceptions(content)
     assert_equal_content(content_new, expected)
