+++ scripts/testng-junit/src/tests/test_migrate_inject_constructor.py
@@ -1,4 +1,4 @@
-from setup import testng2junit
+from setup import testng2junit4
 from setup import assert_equal_content
 
 content = """
@@ -37,5 +37,5 @@ expected = """
 
 
 def test_migrate_inject_constructor():
-    content_new = testng2junit.migrate_inject_constructor('SomeClass', content)
+    content_new = testng2junit4.migrate_inject_constructor('SomeClass', content)
     assert_equal_content(content_new, expected)
