+++ scripts/testng-junit/src/tests/test_migrate_listeners.py
@@ -1,4 +1,4 @@
-from setup import testng2junit
+from setup import testng2junit4
 from setup import assert_equal_content
 
 content = '''
@@ -29,5 +29,5 @@ public class SomeTest
 }
 '''
 def test_migrate_listeners():
-    content_new = testng2junit.migrate_listeners(content)
+    content_new = testng2junit4.migrate_listeners(content)
     assert_equal_content(content_new, expected)
